third_party/blink/web_tests/platform/generic/fast/loader/invalid-module-specifier-expected.txt

CONSOLE ERROR: Uncaught TypeError: Failed to resolve module specifier "specifier-without-valid-prefix". Relative references must start with either "/", "./", or "../".
CONSOLE ERROR: Uncaught TypeError: Failed to resolve module specifier "./foo". Invalid relative url or base scheme isn't hierarchical.
